requirements.txt
Flask==3.0.2
honeybadger==0.19.0
python-dotenv==1.0.1
transformers==4.39.0
torch
sentencepiece